requirements.txt
fastapi[all]
uvicorn[standard]
pymongo==3.11.3
pydantic
requests==2.26.0
passlib==1.7.4
py-bcrypt==0.4
pyjwt==2.2.0
python-jose[cryptography]==3.3.0
jinja2==3.0.2